добавлены кубы и расстояние в 3D

+++ Csharp/Семинары/Занятие/расстояние в 2D/Program.cs
@@ -4,5 +4,5 @@ int y1 = 6;
 int x2 = 2;
 int y2 = 1;
 
-int length = (int)Math.Sqrt(Math.Pow(x1-x2,2) + Math.Pow(y1-y2,2));
+double length = Math.Sqrt(Math.Pow(x1-x2,2) + Math.Pow(y1-y2,2));
 System.Console.WriteLine(length);
